fix: reconcile AGENTS.md, ARCHITECTURE.md, README.md against code + add distributed test infra
- Make DownloadTask.download_dir and downloader_id non-nullable (match spec)
- Change downloader_id FK from SET NULL to RESTRICT, cascade-delete tasks on downloader removal
- Change Agent.llm_enabled default from False to True
- Add updated_at to FileResource/PendingDecision in AGENTS.md
- Document missing defaults in AGENTS.md (llm_enabled, scope_channel_wide, DownloaderInstance.status)
- Add TVSeries relationships (raw_title_mappings, pending_decisions) to AGENTS.md
- Add missing env vars (QUEUE_BACKEND, LLM_ENABLE_THINKING, DEBUG, LOG_LEVEL, etc.) to README.md
- Fix _extract_download_urls to detect .torrent links in entry link field (Nyaa.si compat)
- Fix file_size schema coercion float→int
- Add utcnow() helper for naive UTC datetimes (PostgreSQL compatibility)
- Change file_size column from Integer to BigInteger (prevents PostgreSQL int32 overflow)
- Add asyncpg dependency for PostgreSQL support
- Split integration tests into docker-compose.test-all-in-one.yml (SQLite+MemoryQueue) and docker-compose.test-distributed.yml (PostgreSQL+Redis+2 replicas)
- Fix 5 tests for non-nullable constraints

+++ app/models/file_resource.py
@@ -3,7 +3,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, UniqueConstraint, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -35,7 +35,7 @@ class FileResource(Base):
     audio_codec: Mapped[str | None] = mapped_column(String(100), nullable=True)
     subtitle_type: Mapped[str | None] = mapped_column(String(100), nullable=True)
     container: Mapped[str | None] = mapped_column(String(20), nullable=True)
-    file_size: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    file_size: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     torrent_url: Mapped[str] = mapped_column(String(2048), nullable=False)
     detail_url: Mapped[str | None] = mapped_column(String(2048), nullable=True)
     published_at: Mapped[datetime | None] = mapped_column(DateTime, nullable=True)
